202509_AITinkerers/pyproject.toml
[project]
name = "202509-aitinkerers"
version = "0.1.0"
description = "Personal Diet and Nutrition Coach and Assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elevenlabs>=2.16.0",
    "mirascope>=1.25.6",
    "pydantic-ai>=1.0.10",
    "fastapi>=0.115.0",
    "streamlit>=1.38.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "mcp>=0.5.0",
    "pydantic-ai-slim[anthropic]>=1.0.10",
    "chainlit>=1.1.0",
]

[tool.uv.workspace]
members = ["server"]
